Fix blockDifficulty interpreting in the ExternalStateForFvm interface

+++ modFastVM/test/org/aion/vm/ExecutionContextUnitTest.java
@@ -297,6 +297,6 @@ public class ExecutionContextUnitTest {
         end += FvmDataWord.SIZE;
         assertEquals(
                 context.getBlockDifficulty(),
-                FvmDataWord.fromBytes(Arrays.copyOfRange(encoding, start, end)).toLong());
+                FvmDataWord.fromBytes(Arrays.copyOfRange(encoding, start, end)));
     }
 }
